Develop add mosaic support

+++ georeference/jobs/actions/create_geo_services_test.py
@@ -14,23 +14,23 @@ from .create_geo_services import run_process_geo_services
 # Initialize the logger
 LOGGER = logging.getLogger(__name__)
 
-def test_runProcessGeoServices_success():
+def test_run_process_geo_services_success():
     """ The proper working of the action for processing geo services for a geo images. """
     try:
         # Perform the test
-        testData = create_test_data('test_runProcessGeoServices_success')
-        testMapfile = os.path.join(os.path.dirname(os.path.realpath(__file__)),
-                           '../../georeference_tests/data_output/%s.map' % 'test_runProcessGeoServices_success')
-        pathGeoImage = run_process_geo_image(
-            testData['transformationObj'],
-            testData['srcPath'],
-            testData['trgPath'],
+        test_data = create_test_data('test_runProcessGeoServices_success')
+        test_mapfile = os.path.join(os.path.dirname(os.path.realpath(__file__)),
+                           '../../data_output/%s.map' % 'test_runProcessGeoServices_success')
+        path_geo_image = run_process_geo_image(
+            test_data['transformationObj'],
+            test_data['srcPath'],
+            test_data['trgPath'],
             logger=LOGGER
         )
 
         subject = run_process_geo_services(
-            testMapfile,
-            pathGeoImage,
+            test_mapfile,
+            path_geo_image,
             "test_runProcessGeoServices_success",
             "test_runProcessGeoServices_success",
             "Test",
@@ -38,9 +38,10 @@ def test_runProcessGeoServices_success():
             with_wcs = True
         )
 
-        assert True == os.path.exists(subject)
+        assert subject is not None
+        assert os.path.exists(subject)
     finally:
-        if os.path.exists(pathGeoImage):
-            os.remove(pathGeoImage)
-        if os.path.exists(subject):
+        if path_geo_image is not None and os.path.exists(path_geo_image):
+            os.remove(path_geo_image)
+        if subject is not None and os.path.exists(subject):
             os.remove(subject)
